requirements.txt
blinker==1.9.0
click==8.3.3
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==4.17.0
python-dotenv==1.2.2
Werkzeug==3.1.8
